+++ JooBank/src/main/java/page/DepositPageController.java
@@ -15,11 +15,11 @@ public class DepositPageController implements Controller {
 		String ac_number = request.getParameter("ac_number");
 		
 //		AccountVO vo = new AccountVO();
-//		vo.setAc_number(Integer.parseInt(ac_number));
+//		vo.setAc_number(Long.parseLong(ac_number));
 		
 		AccountDAO dao = new AccountDAO();
 //		AccountVO account = dao.getAc_number(vo);
-		AccountVO account = dao.getAccount(Integer.parseInt(ac_number));
+		AccountVO account = dao.getAccount(Long.parseLong(ac_number));
 		
 		request.setAttribute("account", account);
